+++ wiliwili/include/api/bilibili/result/mine_later_result.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "nlohmann/json.hpp"
+#include "bilibili/util/json.hpp"
 
 namespace bilibili {
 
@@ -27,10 +27,10 @@ public:
     Up owner;
     int duration;  // video length in seconds
     std::string title;
-    int aid;
+    uint64_t aid;
     Stat stat;
     std::string bvid;
-    int cid;
+    uint64_t cid;
 };
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE(WatchLaterItem, aid, pic, title, duration, owner, stat, bvid, cid);
